Remove zimCheck.h
There is no need to push all those function in a header file.
